+++ src/index.d.ts
@@ -16,8 +16,6 @@ declare module "bizcharts" {
   export import PathUtil = _BizCharts.PathUtil;
   export import track = _BizCharts.track;
   export import setTheme = _BizCharts.setTheme;
-  export import AlignXType = _BizCharts.AlignXType;
-  export import AlignYType = _BizCharts.AlignYType;
   export import Axis = _BizCharts.Axis;
   export import AxisProps = _BizCharts.AxisProps;
   export import Base = _BizCharts.Base;
@@ -25,23 +23,17 @@ declare module "bizcharts" {
   export import Chart = _BizCharts.Chart;
   export import ChartProps = _BizCharts.ChartProps;
   export import Coord = _BizCharts.Coord;
-  export import CoordProps = _BizCharts.CoordProps;
-  export import CoordType = _BizCharts.CoordType;
-  export import CrosshairsType = _BizCharts.CrosshairsType;
+  export import CoordProps = _BizCharts.ChartProps;
   export import Facet = _BizCharts.Facet;
   export import FacetProps = _BizCharts.FacetProps;
-  export import FacetType = _BizCharts.FacetType;
   export import Geom = _BizCharts.Geom;
   export import GeomProps = _BizCharts.GeomProps;
-  export import GeomType = _BizCharts.GeomType;
   export import Guide = _BizCharts.Guide;
   export import GuideProps = _BizCharts.GuideProps;
   export import Label = _BizCharts.Label;
   export import LabelProps = _BizCharts.LabelProps;
   export import Legend = _BizCharts.Legend;
-  export import LegendProps = _BizCharts.LegendProps;
-  export import MarkerType = _BizCharts.MarkerType;
-  export import PositionType = _BizCharts.PositionType;
+  export import LegendProps = _BizCharts.LabelProps;
   export import Tooltip = _BizCharts.Tooltip;
   export import TooltipProps = _BizCharts.TooltipProps;
   export import View = _BizCharts.View;
@@ -99,22 +91,15 @@ declare namespace _BizCharts{
 
   // some config type
   export let AxisTile: G2.AxisTile;
-  export type AlignXType = 'left' | 'middle' | 'right';
-  export type AlignYType = 'top' | 'middle' | 'bottom';
-  export type CoordType = 'rect' | 'polar' | 'theta' | 'helix';
-  export type CrosshairsType = 'rect' | 'x' | 'y' | 'cross';
-  export type FacetType = 'rect' | 'list' | 'circle' | 'tree' | 'mirror' | 'matrix';
   export type GeomType = 'point' | 'path' | 'line' | 'area' | 'interval' | 'polygon' | 'edge' | 'schema' | 'heatmap' | 'pointStack' | 'pointJitter' | 'pointDodge' | 'intervalStack' | 'intervalDodge' | 'intervalSymmetric' | 'areaStack' | 'schemaDodge';
   export type MarkerType = 'circle' | 'square' | 'bowtie' | 'diamond' | 'hexagon' | 'triangle' | 'triangle-down' | 'hollowCircle' | 'hollowSquare' | 'hollowBowtie' | 'hollowDiamond' | 'hollowHexagon' | 'hollowTriangle' | 'hollowTriangle-down' | 'cross' | 'tick' | 'plus' | 'hyphen' | 'line';
-  export type PositionType = 'top' | 'bottom' | 'left' | 'right';
-  
   /**
    * components
    */
   export interface AxisProps extends React.Props<any> {
     name?: string;
     visible?: boolean;
-    position?: PositionType;
+    position?: 'top' | 'bottom' | 'left' | 'right';
     title?: object;
     line?: G2.Styles.line;
     tickLine?: G2.Styles.tickLine;
@@ -187,8 +172,8 @@ declare namespace _BizCharts{
   }
 
   export interface CoordProps extends React.Props<any> {
-    type?: CoordType,
-    rotate?: number,
+    type?: 'rect' | 'polar' | 'theta' | 'helix',
+    rotate?: 'number',
     scale?: [number, number],
     reflect?: 'x' | 'y',
     radius?: number,
@@ -198,7 +183,7 @@ declare namespace _BizCharts{
   }
 
   export interface FacetProps extends React.Props<any> {
-    type?: FacetType;
+    type?:  'rect' | 'list' | 'circle' | 'tree' | 'mirror' | 'matrix';
     fields?: string | any[];
     margin?: number | number[];
     padding?: number | number[];
@@ -249,7 +234,7 @@ declare namespace _BizCharts{
   export interface LegendProps extends React.Props<any> {
     name?: string;
     visible?: string;
-    position?: PositionType;
+    position?: 'top' | 'left' | 'right' | 'bottom';
     title?: boolean;
     offsetX?: number;
     offsetY?: number;
@@ -289,7 +274,7 @@ declare namespace _BizCharts{
   export interface TooltipProps extends React.Props<any> {
     showTitle?: boolean;
     crosshairs?: {
-      type?: CrosshairsType;
+      type?: 'rect' | 'x' | 'y' | 'cross';
       style?: G2.Styles.background | G2.Styles.line;
     };
     offset?: number;
@@ -303,7 +288,7 @@ declare namespace _BizCharts{
     inPlot?: boolean;
     follow?: boolean;
     shared?: boolean;
-    position?: PositionType;
+    position?: 'top' | 'bottom' | 'left' | 'right';
 
   }
 
@@ -383,8 +368,8 @@ declare namespace _BizCharts{
 
     interface HtmlProps {
       position?: object | any[] | ((xScale?: any, yScale?: any) => any);
-      alignX?: AlignXType;
-      alignY?: AlignYType;
+      alignX?: 'left' | 'middle' | 'right';
+      alignY?: 'top' | 'middle' | 'bottom';
       html?: string;
       zIndex?: number;
       offsetX?: number;